pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "5hintaro"},
]
dependencies = [
    "pyyaml>=6.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}
